services/agent-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
prisma>=0.15.0
redis>=5.2.1
asyncpg>=0.30.0
python-dotenv>=1.0.1

# LLM Provider SDKs
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.34.0

# Testing and development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0